Address all errors and build issues on Windows.

+++ src/vm/math/Mathmod.cpp
@@ -13,6 +13,7 @@
 #endif
 
 #include <cfenv>
+#include <algorithm>
 
 #include <debug.h>
 #include "Mathmod.h"
@@ -439,7 +440,7 @@ Math_export_int(vm_registers_t &r, vm_t &vm)
     assert(values->get_element_type()->is_equal(intrinsic_type_desc::type<word_t>().get()));
     const auto values_len = values->get_length();
 
-    if (buffer->get_length() != (sizeof(word_t) * values_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(word_t) * values_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Insert as big endian
@@ -467,7 +468,7 @@ Math_export_real(vm_registers_t &r, vm_t &vm)
     assert(values->get_element_type()->is_equal(intrinsic_type_desc::type<real_t>().get()));
     const auto values_len = values->get_length();
 
-    if (buffer->get_length() != (sizeof(real_t) * values_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(real_t) * values_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Insert as big endian
@@ -502,7 +503,7 @@ Math_export_real32(vm_registers_t &r, vm_t &vm)
     assert(values->get_element_type()->is_equal(intrinsic_type_desc::type<real_t>().get()));
     const auto values_len = values->get_length();
 
-    if (buffer->get_length() != (sizeof(float) * values_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(float) * values_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Insert as big endian
@@ -712,7 +713,7 @@ Math_import_int(vm_registers_t &r, vm_t &vm)
     assert(result->get_element_type()->is_equal(intrinsic_type_desc::type<word_t>().get()));
     const auto result_len = result->get_length();
 
-    if (buffer->get_length() != (sizeof(word_t) * result_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(word_t) * result_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Process as big endian
@@ -739,7 +740,7 @@ Math_import_real(vm_registers_t &r, vm_t &vm)
     assert(result->get_element_type()->is_equal(intrinsic_type_desc::type<real_t>().get()));
     const auto result_len = result->get_length();
 
-    if (buffer->get_length() != (sizeof(real_t) * result_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(real_t) * result_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Process as big endian
@@ -769,7 +770,7 @@ Math_import_real32(vm_registers_t &r, vm_t &vm)
     assert(result->get_element_type()->is_equal(intrinsic_type_desc::type<real_t>().get()));
     const auto result_len = result->get_length();
 
-    if (buffer->get_length() != (sizeof(float) * result_len))
+    if (static_cast<size_t>(buffer->get_length()) != (sizeof(float) * result_len))
         throw vm_user_exception{ "Invalid buffer size based on conversion" };
 
     // Process as big endian
